requirements.txt
amightygirl.paapi5-python-sdk==1.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
emoji-country-flag==1.3.2
holidays==0.40
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==9.5.0
pyTelegramBotAPI==4.14.1
python-amazon-paapi==5.0.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==1.26.6